pyproject.toml
[project]
name = "netbox-dio"
version = "0.1.0"
description = "A high-level Python wrapper for the NetBox Diode SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "netboxlabs-diode-sdk>=0.0.1",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=26.0.0",
    "ruff>=0.15.0",
    "isort>=8.0.0",
    "mypy>=1.20.0",
    "typer>=0.24.0",
    "pyyaml>=6.0.0",
    "pytest-mock>=3.10.0",
]
docs = [
    "mkdocs>=1.4.0",
    "mkdocs-material>=9.0.0",
    "pdoc>=14.0.0",
    "mkdocs-macros-plugin>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
netbox-dio = "netbox_dio.cli.app:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-x --strict-markers"
markers = [
    "unit: unit tests",
    "integration: integration tests",
]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
warn_return_types = true
warn_unused_configs = true
